Fix: discountType 타입 변경

+++ backend/src/main/java/com/ssafy/eureka/domain/card/dto/response/CardDetailBenefitList.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class CardDetailBenefitList {
 
     String discountCostType;
-    float discountCost;
+    double discountCost;
     String largeCategoryName;
 
 }
